// code/validation/error-types.ts
export enum ErrorTypes {
  TABLEAU_ATTENDU = "TABLEAU_ATTENDU",
  CHAMP_REQUIS = "CHAMP_REQUIS",
  ATTRIBUT_INVALIDE = "ATTRIBUT_INVALIDE",
  DATE_INVALIDE = "DATE_INVALIDE",
  CHIFFRES_INVALIDES = "CHIFFRES_INVALIDES",
  CHIFFRES_INVALIDE_4 = "CHIFFRES_INVALIDE_4",
  CHIFFRES_INVALIDE_5 = "CHIFFRES_INVALIDE_5",
  CHIFFRES_INVALIDE_6 = "CHIFFRES_INVALIDE_6",
  CHIFFRES_INVALIDE_14 = "CHIFFRES_INVALIDE_14",
  CHARACTERES_MAJUSCULE_3_INVALIDE = "CHARACTERES_MAJUSCULE_3_INVALIDE",
  TYPE_INVALIDE = "TYPE_INVALIDE",
  PATRONYME_ERREUR = "PATRONYME_ERREUR",
  EMAIL_ERREUR = "EMAIL_ERREUR"
}